1.footer corrected 2.gird 3*3 corrected 3.removed search icon 4.searchbar name changed to promtbar
