+++ .prettierrc
@@ -1,8 +1,10 @@
 {
-    "semi": false,
+    "semi": true,
     "trailingComma": "all",
     "singleQuote": true,
     "printWidth": 80,
     "tabWidth": 2,
+   " bracketSpacing": true,
     "plugins": ["prettier-plugin-tailwindcss"]
   }
+
