+++ app/components/footer/footer.tsx
@@ -2,7 +2,7 @@ import React from 'react'
 import Link from 'next/link'
 const Footer = () => {
   return (
-    <div className="footer flex min-h-full flex-col ">
+    <div className="footer flex min-h-full flex-col flex-shrink-0">
       <div className="mt-10 ">
         <footer className="bg-gray-800 text-center lg:text-left">
           <div className="p-4 text-center text-neutral-700 dark:text-neutral-200">
