+++ app/globals.css
@@ -6,16 +6,6 @@ body {
   background: #fff;
   color: #000;
 }
-.main-container {
-  display: flex;
-  flex-direction: column;
-  min-height: 100vh;
-}
 
-.layout-container {
-  flex: 1;
-}
 
-.footer {
-  flex-shrink: 0;
-}
+
