+++ app/login/page.tsx
@@ -1,8 +1,10 @@
 import React from 'react'
 import LoginForm from '../components/login/login'
+import Layout from '../components/layout'
 const LoginPage = () => {
   return (
     <div>
+      
       <LoginForm/>
     </div>
   )
