+++ app/signup/page.tsx
@@ -1,8 +1,11 @@
 import React from 'react'
 import SignupForm from '../components/signup/signup'
+import Layout from '../components/layout'
+Layout
 const SignupPage = () => {
   return (
     <div>
+      
       <SignupForm/>
     </div>
   )
